Deckクラスを分離できる
