+++ src/game21.py
@@ -1,6 +1,5 @@
 from effect import Effect
-import random
-from collections import Counter
+from deck import Deck
 from calculate_burst_probability import CalculateBurstProbability
 import joblib
 import pandas as pd
@@ -46,40 +45,6 @@ class Player:
             Effect.highlight_line("山札が尽きました！！")
             Effect.display_with_pause()
             return False
-
-
-class Deck:
-    def shuffle(self):
-        random.shuffle(self.cards)
-
-    def draw(self):
-        return self.cards.pop() if self.cards else None
-
-    def reset(self):
-        self.cards = self.initial_deck_cards()
-        self.shuffle()
-
-    def initial_deck_cards(self):
-        return [i for i in range(1, 12)]
-
-    def identify_remaining_cards(self, player_hand, opponent_hand):
-        card_counter = Counter(self.initial_deck_cards())
-
-        for card in player_hand:
-            if card != "?":
-                card_counter[card] -= 1
-
-        for card in opponent_hand:
-            if card != "?":
-                card_counter[card] -= 1
-
-        remaining_cards = [
-            card
-            for card, count in card_counter.items()
-            if count > 0
-            for _ in range(count)
-        ]
-        return remaining_cards
 
 
 class Game21:
